+++ example/podcast_search_example.dart
@@ -4,7 +4,7 @@ void main() async {
   var search = Search();
 
   /// Search for the "It's a Widget" podcast.
-  var results = await search.search('widgets', country: Country.UNITED_KINGDOM, limit: 10);
+  var results = await search.search('widgets', country: Country.unitedKingdom, limit: 10);
 
   /// List the name of each podcast found.
   results.items.forEach((result) {
@@ -20,7 +20,7 @@ void main() async {
   });
 
   /// Find the top 10 comedy podcasts in the UK.
-  var charts = await search.charts(genre: 'Education', limit: 10, country: Country.UNITED_KINGDOM);
+  var charts = await search.charts(genre: 'Education', limit: 10, country: Country.unitedKingdom);
 
   /// List the name of each podcast found.
   charts.items.forEach((result) {