+++ lib/src/search/podcast_index_search.dart
@@ -199,9 +199,9 @@ class PodcastIndexSearch extends BaseSearch {
   @override
   Future<SearchResult> search(
       {String? term,
-      Country country = Country.NONE,
-      Attribute attribute = Attribute.NONE,
-      Language language = Language.NONE,
+      Country country = Country.none,
+      Attribute attribute = Attribute.none,
+      Language language = Language.none,
       int limit = 0,
       int version = 0,
       bool explicit = false,
@@ -223,7 +223,7 @@ class PodcastIndexSearch extends BaseSearch {
 
   /// Fetches the list of top podcasts
   /// Optionally takes a [limit] and [Country] filter. Defaults to
-  /// limit of 20 and the UK.
+  /// limit of 20 and no specified country.
   ///
   /// The charts is returned as a 'feed'. In order to be compatible with
   /// [SearchResult] we need to parse this feed and fetch the underlying
@@ -232,7 +232,7 @@ class PodcastIndexSearch extends BaseSearch {
   /// cache the results.
   @override
   Future<SearchResult> charts(
-      {Country country = Country.UNITED_KINGDOM,
+      {Country country = Country.none,
       int limit = 20,
       bool explicit = false,
       String genre = '',