+++ lib/src/search/search.dart
@@ -13,20 +13,20 @@ class Search {
   String _term = '';
 
   /// If this property is not-null search results will be limited to this country
-  Country _country = Country.NONE;
+  Country _country = Country.none;
 
   /// If this property is not-null search results will be limited to this genre
   String _genre = '';
 
   /// By default, searches will be performed against keyword(s) in [_term].
   /// Set this property to search against a different attribute.
-  Attribute _attribute = Attribute.NONE;
+  Attribute _attribute = Attribute.none;
 
   /// Limit the number of results to [_limit]. If zero no limit will be applied
   int _limit = 0;
 
   /// If non-null, the results will be limited to the language specified.
-  Language _language = Language.NONE;
+  Language _language = Language.none;
 
   /// Set to true to disable the explicit filter.
   bool _explicit = false;
@@ -52,9 +52,9 @@ class Search {
   /// value to search by a different attribute such as Author, genre etc.
   Future<SearchResult> search(
     String term, {
-    Country country = Country.NONE,
-    Attribute attribute = Attribute.NONE,
-    Language language = Language.NONE,
+    Country country = Country.none,
+    Attribute attribute = Attribute.none,
+    Language language = Language.none,
     int limit = 0,
     int version = 0,
     bool explicit = false,
@@ -92,7 +92,7 @@ class Search {
 
   /// Fetches the list of top podcasts
   /// Optionally takes a [limit] and [Country] filter. Defaults to
-  /// limit of 20 and the UK.
+  /// limit of 20 and no specified country.
   ///
   /// The charts is returned as a 'feed'. In order to be compatible with
   /// [SearchResult] we need to parse this feed and fetch the underlying
@@ -100,7 +100,7 @@ class Search {
   /// the infrequent update of the chart feed it is recommended that clients
   /// cache the results.
   Future<SearchResult> charts({
-    Country country = Country.UNITED_KINGDOM,
+    Country country = Country.none,
     int limit = 20,
     bool explicit = false,
     String genre = '',